processor/requirements.txt
rembg==2.0.50
redis==4.6.0
Pillow==10.1.0
numpy==1.24.3 